+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 -----
 - Removed default value at translation method
 - Removed bootstrap file
+- Removed default parameter
+- Added replacing parameters in translation strings
 
 0.0.1
 -----